Add files via upload
Recognizing multi-digit numbers in photographs captured at street level is an important
component of modern-day map making. A classic example of a corpus of such street
level photographs is Google’s Street View imagery comprised of hundreds of millions of
geo-located 360 degree panoramic images. The ability to automatically transcribe an
address number from a geo-located patch of pixels and associate the transcribed
number with a known street address helps pinpoint.